[MyOrders] Backend: Order: OrderNumber
Added Domain Service for generating OrderNumber
OrderNumber now is generated by server
Updated rest file

+++ src/Backend/MyOrders.Application/Services/AddressService.cs
@@ -37,6 +37,12 @@ namespace MyOrders.Application.Services
         public async Task DeleteAsync(int id)
         {
             var address = await GetAddressAsync(id);
+            
+            if (address.Customer is not null)
+            {
+                throw new BusinessException($"Cannot delete Address with id: '{address.Id}', because is used by Customer with id: '{address.Customer.Id}'");
+            }
+
             await _addressRepository.DeleteAsync(address);
         }
